+++ hw/bsp/ruuvitag_rev_b/boot-nrf52xxaa.ld
