+++ hw/bsp/ruuvitag_rev_b/nrf52xxaa.ld
