+++ hw/bsp/ruuvitag_rev_b/ruuvitag_rev_b_download.cmd
