+++ hw/bsp/ruuvitag_rev_b/split-ruuvitag_rev_b.ld
